Rewrote methods in entities without try-catch block

+++ app/entities/users/index.js
@@ -10,15 +10,34 @@ async function getUserById(id) {
 }
 
 async function deleteUserById(id) {
+  const result = await User.query().findById(id);
+  if (!result) {
+    throw ErrorService.errorThrow(404);
+  }
   await User.query().delete().findById(id);
 }
 
 async function addUser(params) {
+  const result = await User.query().findById(params.id);
+  if (result) {
+    throw ErrorService.errorThrow(400);
+  }
   await User.query().insert(params);
 }
 
 async function updateUserById(id, params) {
-  await User.query().update(params).findById(id);
+  const result = User.query().findById(id);
+  if (!result) {
+    throw ErrorService.errorThrow(404);
+  }
+  await User.query()
+    .update({
+      name: `${params.name}`,
+      surname: `${params.surname}`,
+      email: `${params.email}`,
+      password: `${params.password}`,
+    })
+    .findById(id);
 }
 
 module.exports = {
